Split workspace title menu content
